requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Pillow==10.1.0
matplotlib==3.8.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
sentence-transformers==2.2.2
open-clip-torch==2.23.0
Flask-Cors==4.0.0
